export/pyproject.toml
[project]
name = "customer-daisy"
version = "1.0.0"
description = "DaisySMS & Customer Creation System with Mail.tm Integration"
requires-python = ">=3.8"
dependencies = [
    "rich>=13.7.0",
    "requests>=2.31.0", 
    "faker>=19.6.2",
    "python-dateutil>=2.8.2",
    "click>=8.1.7",
    "tabulate>=0.9.0",
    "colorama>=0.4.6",
    "pyperclip>=1.8.2",
    "questionary>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
customer-daisy = "main:main"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=7.0.0",
]